Convert Heuristic from int to double

+++ src/main/java/SlayTheSpireAIMod/AIs/CombatAIs/AbstractCAI.java
@@ -204,10 +204,10 @@ public abstract class AbstractCAI {
             HashSet<CardSequence> ended = getDistantPossibilities();
             logger.info("Current State: " + this.toString());
             logger.info("Number of Distant Pos: " + ended.size());
-            int bestEval = h.evaluation(this);
+            double bestEval = h.evaluation(this);
             CardSequence bestState = this;
             for(CardSequence state : ended){
-                int eval = h.evaluation(state);
+                double eval = h.evaluation(state);
                 if(eval < bestEval){
                     bestEval = eval;
                     bestState = state;
